Remove LGPL from permitted licenses list
This dependency decision goes back to when we first added the
LicenseFinder gem:
https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/3775

This is probably because we were using Sidekiq (which is licensed under
the LGPL) from before we added the LicenseFinder job:
https://gitlab.com/gitlab-org/gitlab-foss/-/commit/c7bb3a1f

However, the license is currently on our unacceptable licenses list:
https://about.gitlab.com/handbook/engineering/open-source/#unacceptable-licenses

The existing LicenseFinder configuration allowed any new LGPL
dependencies to pass CI with no issues. With this change, we will
require individual dependencies that use the LGPL to be approved
one-by-one, rather than as a blanket approval for the license itself.

GitLab-internal reference:
https://gitlab.com/gitlab-com/legal-and-compliance/-/issues/672