Manually approve existing LGPL dependencies
As a matter of general policy, we don't want to add new LGPL
dependencies. However, we were already using these ones, and Legal have
confirmed it's OK to keep them:
https://gitlab.com/gitlab-com/legal-and-compliance/-/issues/672

So here we allow them on the package level - rather than the license
level - to make it clearer that while these LGPL dependencies are
permitted, we should not be adding new ones without seeking approval
from Legal first.