Sleep instead of scheduling in the future
When handling Sidekiq workers that use replicas, we add a sleep in the
worker when the minimum delay is not yet reached. This replaces the
strategy we do now where we schedule jobs 1 second in the future.

We are testing this because Sidekiq scheduled set processing can be
expensive and the latencies are high.