Refactor assignee and reviewer state to allow nil values
Refactors all the `attention_required` variable names to be
`attention_requested` to better match the feature name.

Changes the AttentionRequest service to be a ToggleAttentionRequested
service which will toggle between `attention_requested` and `nil`

This also hooks it up all on the frontend so that the assignees
and reviewers in the sidebar send the correct GraphQL mutation
when the attention request button is clicked.

Changelog: changed

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/343325