GithubImporter: Fix import duration metric
Since `project.full_path` is user provided information, this means that
the cardinality of this metric is not bound and we can't initialize the
metric properly. Because the metric isn't initialized and the project is
included we can never track the rate past 1 for each counter in the
histogram: A histogram is a series of counters, in total there are
`number of buckets * number of label combinations` counters that
Prometheus uses to calculate the quantiles. For that, we're using a
[rate](https://prometheus.io/docs/prometheus/latest/querying/functions/#rate),
which is the increase of a counter in seconds.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/341958

Changelog: fixed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/74610