Restore a rescue in PreventCrossDatabaseModification
This was removed in 5ec78473b9d26d76271e3b911ac6ec8edb41511e and I had
to add it back because we can't raise exceptions in development for all
cross-modifications because so far we've only ensured the allowlist is
correct for specs files and don't have inline allows for all the known
violations. I was able to see a failure by creating a pipeline with
`environment` and it failed with 500. Since we want to be able to enable
the feature flag and env var by default we don't want to break
everyone's local development experience.