Fix discussions filter not working
This bug was introduced by
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/69933 when we
changed the mutation to append or update discussions.

When changing the discussions filter, we must clear the existing
discussions before fetching the filtered list.

This also disables the filter when discussions are loading to prevent
race conditions where the response of a previous request could show up
if the filters are switched very quickly

Changelog: fixed