Fixes cross-db modification for Ci::JobArtifacts::DestroyBatchService
Split out main, and geo DB out of CI transaction

We should not run main, and geo DB operations in the context of a CI
transaction. We split it out into a pre-, and post- set of DB
operations.

At this point it should be OK for the security_finding to be deleted,
regardless of whether the Ci::JobArtifact are successfully delete or
not. If the Ci::JobArtifact deletion fails, another worker will try
again.

Changelog: changed