Temporarily set the RspecFlaky::Report::OUTDATED_DAYS_THRESHOLD to 30
This is needed since we're trying to reset the report to a good known
state, which is older than 7 days. Without this change, most flaky
specs would be removed since their last flaky time is older than 7 days.

Once the report is refreshed properly (we expect around 170 flaky
tests), we can revert this change.

Signed-off-by: Rémy Coutable <remy@rymai.me>