Remove support for sticking to old and new keys
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/67773 we added
support for multiple databases in the load balancer. This was released
in 14.4. As part of these changes, we added support for reading/writing
to both the old and new Redis keys, making it easier to switch between
the old and new version of the code.

In this commit we remove the use of the old key format, as it's no
longer needed.

This fixes https://gitlab.com/gitlab-org/gitlab/-/issues/345208

Changelog: changed