Simplify sanitizing emojis from localStorage
We currently sanitize all emojis with dompurify when we load them from
localStorage.

See also: https://gitlab.com/gitlab-org/gitlab/-/issues/339345

Given that the emoji should only contain emoji characters, we can simply
verify all of them with a regex. As soon as one of the emojis contains a
non-emoji character, we can simply drop it from the list of emojis.

Changelog: performance