Add GraphQL to the Environment Rollback Components
This adds the necessary local queries and mutations required to trigger
a rollback/re-deploy via the new local GraphQL state management set-up.
This will be used as part of the new environments page. It pulls the use
of isLastDeployment out of the environment object as well, and relies on
it being passed in as a prop.