Make Environment Stop work with GraphQL
The new environments page is built using local GraphQL state management,
and so the ability to stop environments needs to rely on it instead of
the event hub.

To ensure everything still works until I am ready to delete the old
environments page, the new opt-in graphql prop is added to enable this
functionality.