Remove unused cluster runner creation code causing cross-modification
As discovered at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/74735#note_740614507
this code causes a CrossDatabaseModificationAcrossUnsupportedTablesError
since it writes to ci and non-ci tables in the context of a single
transaction. You can read more at
https://docs.gitlab.com/ee/development/database/multiple_databases.html#removing-cross-database-transactions .

This code is apparently not used and not needed anymore as we don't
allow people to install runners this way anymore. For now I decided to
just focus on removing the small amount of code that was causing
problems but it seems there could be a deeper rabbit hole here of
removing lots of unused code but I decided not to tackle that all at
once.