Add new environment item component
A new environment component is needed to migrate to the newly
recommended UX. It is currently behind the feature flag
:new_environments_table, and relies on Apollo for local state
management.

Most of the internal computed values and functions are borrowed from the
old environments component
app/assets/javascripts/environments/components/environment_item.vue,
which will be removed once the rollout to the new UI is complete.