Generate site profiles with set URLs
This updates the DAST site profile mocks to use pre-defined URLs instead
of the auto-generated ones. We need this because auto-generated URLs
dependened on the entity's ID, which could be different depending on the
environement fixtures are being generated in. This caused some snapshot
failures on master. Making our fixtures more deterministic should
prevent the failure from happening again.