Mark merge request as unchecked when MR is prepared for mergeability
We currently only require a pipeline to be created (if there's any)
before we mark a merge request to be ready for mergeability check.

Before this fix, we are also waiting for other stuff to be done
before we set a merge request ready for mergeability.

In this fix, we split the logic in AfterCreateService so we only
wait for the pipeline to be created. If it's done, we set the
`merge_status` to `unchecked`.

This is behind the `early_prepare_for_mergeability` feature flag.