Expose failed import error through API
Gitlab already have a `projects/:id/import` to expose the import status
of the given project. This endpoint exposes "hard_failures", which are
errors that caused the import to fail.

This commit, changes the Gitlab::Import::ImportFailureService to create
the failure record with the `retry_count: 0`, when it's failing the
import. The `retry_count: 0` is is the condition to expose the failure
as a "hard_failure" in the API.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/346652

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/75440
Changelog: changed