Pass current_user when rendering Markdown in email
When rendering Markdown in notification emails, we pass in the recipient
as current_user in the rendering context so that private references
visible to that user are rendered.

Without this change, only public references are rendered.

This also passes the issuable_reference_expansion_enabled option so that
it is rendered similar to how it is in the web UI.

Changelog: fixed