Rough polish of the optional approver section
Push in some quick and dirty changes. This work is not complete nor
is it fully correct, but I do not want to hold up this work for two
weeks. I will ask the creator to make a follow-up issue for 14.7 for
me to give this page the fuller once-over it truly needs.