Fix spec violations in commit_statuses API request spec
Whenever we create pipelines we need to call `ensure_project_iid!` on
them before saving them otherwise we trigger a write to the
`internal_ids` table in the transaction for saving the model which
causes as cross-database modification.

These cases are all false positives because we already call
`ensure_project_iid!` everywhere we create a pipeline in real
application code.