Add new environment and review app buttons back
The new environments page was lacking the new environment and add review
apps buttons.

It has been tweaked slightly to take advantage of the action primary and
secondary props of the GlTabs, which means I can't take advantage of the
GlModalDirective anymore. Instead, I modify the modal to work via
v-model to show/hide the modal. This means that the adjustment for the
mobile view happens in the GlTabs component now, thereby making the app
component template smaller.