Allow (un)following someone in the user popover
This adds a new button for signed-in users to the user popover to follow
(or unfollow) the user shown in the popover. The goal of this is to
greatly reduce the friction to follow (and unfollow) a user and make the
following feature more widely-adopted. Previously, you had to visit the
profile of the user you are trying to (un)follow and see the button in
the top right of the page.

This also allows you to quickly determine whether or not you are
currently following a person.

Changelog: added