Destroy UpcomingReconciliation when none upcoming
Previously, when a user cancelled their subscription, we would still
show the upcoming reconciliation alert to the user at the time they
would have originally been reconciled. This is because we create the
UpcomingReconciliation record but we never clean it up when the
reconciliation won't actually happen.

This MR destroys the UpcomingReconciliation record once we stop
receiving the next reconciliation date from the Seat Link API.

Issue: https://gitlab.com/gitlab-org/customers-gitlab-com/-/issues/3722