Create bridge vue app
This creates a separate vue app for the bridge show page.
The app is separated from the existing jobs page to keep
the separation of data and behaviors clean on the frontend
and because it will use graphQL in the future to implement
trigger retry actions (job vue app uses REST and vuex).