Add tabs to new environments page
The tabs select the scope of environments to show, and are synced to the
URL query parameters.

Testing the tab selection is a little tricky due to the reactive poll
interval, so if there is no interval, just skip polling.

This is also good in the event polling is disabled.