Fix SMIME signed email sending, do not overwrite Content-Disposition
If signed email sending was enabled the Content-Disposition was set
but empty. This could cause issues with SMTP servers, in particular
sending via Amazon SES did not work at all as they expect a non-empty
Content-Disposition header. Response from SES with empty header was:
  554 Transaction failed: Expected disposition,
  got null (in reply to end of DATA command))

With this fix the Content-Disposition field is no longer overwritten
within the smime_signature_interceptor hook and the mails could be
delivered properly.

Changelog: fixed