Allow external Spamcheck API key to be blank
GitLab's Spamcheck service does not have authentication support on its
own. So, on single-node omnibus-gitlab installations it runs
unauthenticated over localhost. In this scenario, any random text, as
long as it is not blank, will work as an API key. But, this doesn't look
nice from a documentation perspective. Nor does it reflect the current
realit with respect to authentication. Hence, removing the requirement
of an API key to use external Spamcheck service at least until Spamcheck
can do authentication on its own. Once that is the case, we can make
omnibus use authentication for the connection over localhost and turn
this requirement back on.

Changelog: changed
Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>