Add pagination to new environments app
The old page paginates based on the headers returned by the API, and so
we do the same here.

We put the page info header into the GraphQL cache, and fetch it out to
pass it to the pagination component from GitLab UI.

Also the pagination must be wired up to the query parameters to support
deep linking.

I wonder if GlPagination should just have a prop for syncing the query
parameters, like GlTabs does.