Unescape and sanitize protected tag name on create and update
The frontend escapes the names of tags when listing the names in
the dropdown. It also submits the escaped name as the tag to
protect when creating a protected tag.

To fix it, we unescape and sanitize the name so it matches the
appropriate tag name to be protected.

Changelog: fixed