Fix stale Prometheus metrics directory not being restored
metrics_server_spec.rb used a temporary directory to store Prometheus
metrics, but there is a class variable that stores the current
configuration. When the directory was wiped, the configuration was not
properly restored, so any subsequent specs that attempted to log metrics
would fail.

We fix this by restoring the old metrics dir in the config after the
spec runs.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/347499