Generates discussion_id when importing from GitHub
The GitHub importer skips callbacks so the discussion_id isn't
generated. This is fine because we also generate it at runtime when the
note is rendered.

But it is still better to generate it and persist it so that a note does
not have a different discussion id every time it is rendered.

The discussion_id is also going to be used for paginating discussions.