Added support for content level 2 in widget extensions
This creates a new `generateText` method that will correctly
generate the markup required to display each piece of text.
Using this method allows for the extensions to never pass any HTML
through to the base component.
They will instead pass this formatted string and the base
component will use the `generateText` method to create
the correct markup.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/341046