rake: Extract Gitaly checkout into separate task
The way Gitaly is built in source installations, local test, and CI are
slightly different. In a future commit gitlab:gitaly:install will no
longer be used, but we still need to clone and checkout the correct
revision. To do so, this change extracts the cloning and checking out
into a separate Rake task. The install task will depend on that new task
so that functionality remains the same.