test: Non-existing component won't match git SHA
The #component_matches_git_sha? runs a git command to see if the
component is at the correct version. To run it, it uses
Gitlab::Popen, but as a side-effect it creates the directory before it
runs the command. This causes the rake task gitlab:gitaly:clone to
incorrectly determine if a git-clone(1) is needed.

To speed things up, and avoid creation of an empty directory, this
change adds an additional check to see if the component's folder exists
anyway.