Check WAL locations before sleeping
This saves us from the unnecessary sleep if the replica is already up to
date.

Based on previous tests, 50% of the time, the replica is already up to
date so we don't need the sleep.

This also reduces the sleep time because we did not see any stale
replica errors with the previous value. This suggests that the delay may
be too high.