Expose finished_at in job details JSON
This fixes a bug where the time a job has been running (including when
it is complete), the running time, is always labelled **Elapsed time**
in the job sidebar. This, however, is not intended following the
change of https://gitlab.com/gitlab-org/gitlab/-/merge_requests/75811,
and the running time should be labelled **Elapsed time** only when the
job is running and default to the **Duration** text.

The bug is fixed by adding `finished_at` to the `BuildDetailsEntity`
because it was missing there.

Changelog: fixed