Rename projects set for delayed deletion
Projects that are set for delayed deletion
will now have -deleted-:project_id
appended to their name while they are
awaiting deletion.

If they are restored, the path will be reset
unless the path has been taken in the meantime.
In this case, a random alphanumeric string will be
appended to avoid collisions.

Changelog: added