test: Move spawning of Gitaly to GitalySetup
GitalySetup was shelling out to scripts/gitaly-test-spawn to spawn
Gitaly processes, which was quite complicated to understand.

This change simplifies the back-and-forth between the script and the
Ruby file. It avoids writing the PID of the Gitaly processes to file, to
pass them back to the test environment.

With the previous changes included, now scripts/gitaly-test-build &
scripts/gitaly-test-spawn are *only* used in CI. When running tests in a
local GDK, the Ruby code to build and spawn Gitaly in test is directly
called from TestEnv.init.