Add error alerts and badge for webhooks
When a webhook is temporary disabled (due to 5xx errors)
or permanently disabled (due to 4xx errors), we now
show the user an alert (on the edit page) and a badge
(on the index page) to alert the user of what exactly
the issue is.

This commit also extracts all alerts to a shared partial
that will be used on project / group / instance levels.

Changelog: added