Ensure all scanners are translated in alerts
This slightly changes how the scanner names' translations are being
gathered before being passed to the security-scanner-alert component.
By moving all the translation to a shared map, it should harder to
forget to provide a new translation when adding a new scanner.

This is primarily done to translate SAST_IAC to something more
user-friendly.

Changelog: fixed