Add copy issue URL button to vulnerability error message
When creating an issue from a vulnerability, if a second issue is
created for the same vulnerability, an error message is shown. This
commit changes the error message from a haml partial template to a
Vue component, updates its styling to match Storybook, and adds in a
copy issue URL button.

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/76215
EE: true