Add Free Trial dataLayer events for GitLab.org
Related to
https://gitlab.com/gitlab-com/marketing/digital-experience/buyer-experience/-/issues/212,
specifically adds the dataLayer events for /-/trial_registrations/new.

Adds an ops feature flag so we can activate only on GitLab.com,
updates the auth_helper to allow GTM for signed in users, updates
the controller to pass the frontend feature flag, and adds an RSpec test
to verify behavior.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/76305